marker-> place_info
